% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLowPCor.R
\name{RLowPCor}
\alias{RLowPCor}
\title{Relevance low order partial correlation}
\usage{
RLowPCor(data.exp, edgelist, estimator = "pearson", pc.estimator = "shrink")
}
\arguments{
\item{data.exp}{gene expression matrix. Columns are variables and rows are samples.}

\item{edgelist}{edge list. First column are the name of regulators, second coloumn are the target genes and the third column are the edge weights.}

\item{estimator}{a character string indicating which correlation coefficient (or covariance) is to be computed.Options are "pearson", "spearman"
and "kendall". If shrinkage method is used to estimate PC, the estimator is set to "pearson".}

\item{pc.estimator}{a character string indicating which method is used to estimate the PC of nodes connected to shared neigbours. Options are "shrink" and "pc",
correspoinding to the item (c) and (d) in Details: Step 2, respectively.}
}
\value{
a network matrix
}
\description{
Consensus network is built of taking the average ranks of the edges from multiple network predictions.
}
\details{
Step 1:  Extract a sparse and scale-free topology from pre-inferred networks as an indirect edge search space for RLowPCor. Unlike a fully connected network,
the nodes in the sparse network are assumed to connect to their more relevant neighbours. For example, correlation network can be cut with a range of thresholds
until it most fit to scale-free topology. Step 2: Calculate relevance low order partial correlation.  For each pair of nodes connected by an edge in the searching
space, the edge weight is redefined as (a) Pearson correlation if they do not connect to the same set of neighbour nodes, (b) PC by removing shared neighbours and
(c) shrink PC if the covariance matrix used to estimate PC in (b) is not positive definite or invertible. If the searching space is very large, there might still
be a number of irrelevant controls involved in shrink PC procedure (c). An alternative is (d) deleting less connected neighbours of the nodes until the covariance
matrix in (b) is positive definite and invertible.
}
\examples{
##load size 100_1 network DREAM4 datasets
library(networkBMA)
library(RLowPCor)
data(dream4)
data.exp<-dream4ts100[[1]]
#create edge list
edgelist<-dream4gold100[[1]]
edgelist<-edgelist[edgelist[,3]>0,]
##infer RLowPCor network
inf.net=RLowPCor(data.exp = data.exp[,-c(1:2)],edgelist = edgelist)
}
\references{
Sch\"afer J, Strimmer K: A Shrinkage Approach to Large-Scale Covariance Matrix Estimation and Implications for Functional Genomics. Statistical Applications
in Genetics and Molecular Biology, The Berkeley Electronic Press 2005, 4(1).
}

