% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.neighbour.R
\name{shared.neighbour}
\alias{shared.neighbour}
\title{Extract sub-edge list with shared neihbours}
\usage{
shared.neighbour(node1, node2, edgelist)
}
\arguments{
\item{node1}{a character string of the name of candidate node 1}

\item{node2}{a character string of the name of candidate node 2}

\item{edgelist}{the full edge list used to extract the sub-edge list. First column are regulators, second column are target genes and the last column are weigths of the connections.}
}
\value{
Sub-edge list
}
\description{
The function is used to extract the shared neighbours that both connect to a pair of candidate nodes.
}

